requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
openai==0.28.1
httpx==0.28.1
python-dotenv>=1.0.1,<2.0.0
numpy>=1.24.0
pymilvus>=2.3.0
sentence-transformers>=2.2.0 